requirements.txt
fastapi
uvicorn
python-dotenv

langchain
langchain-core
langchain-community

langchain-groq
langchain-mistralai
langchain-openai

groq
mistralai

httpx
pydantic